28/8/2023

+++ Array/Khoang_cach_nho_hon_k.java
@@ -7,20 +7,20 @@ public class Khoang_cach_nho_hon_k {
         while(t-->0){
 
             int n = sc.nextInt();
-            Long k = sc.nextLong();
-            Long arr[] = new Long[n];
+            int k = sc.nextInt();
+            int arr[] = new int[n];
             for(int i = 0; i < n; i++){
-                arr[i] = sc.nextLong();
+                arr[i] = sc.nextInt();
             }
             System.out.println(handle(arr, n, k));
         }
     }
 
-    static int handle(Long arr[], int n, Long k){
+    static int handle(int arr[], int n, int k){
         Arrays.sort(arr);
         int ans = 0;    
         for(int i = 0; i < n; i++){
-            Long val = arr[i]+k;
+            int val = arr[i]+k;
             int y = Arrays.binarySearch(arr, val);
             if(y < 0){  
                 y = -(y + 1);
